+++ src/main/kotlin/SqlCipherApple.kt
@@ -13,7 +13,8 @@ class SqlCipherApple(
     private val srcDir: File,
     private val buildType: BuildType,
     private val config: AppleToolExtension,
-    private val execOperations: ExecOperations
+    private val execOperations: ExecOperations,
+    private val tempOptionString: String
 ) {
     private val amalgamation = SqlcipherExtension.amalgamation
     fun build(opensslIncludeDir: File,
@@ -22,7 +23,8 @@ class SqlCipherApple(
             srcDir,
             runner,
             buildType,
-            execOperations
+            execOperations,
+            tempOptionString
         )
         if (!srcDir.resolve(amalgamation).exists())
             throw GradleException("Amalgamation $amalgamation not found")